+++ frontend/lost/src/containers/Labels/LabelTree.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react'
-// import Graph from 'react-graph-vis'
-import mapTreeToGraph from '../../libs/graph-vis/mapTreeToGraph'
 import EditLabel from './EditLabel'
+import LabelsPage from './LabelsPage'
 
 const options = {
     autoResize: true,
@@ -71,11 +70,13 @@ class LabelTree extends Component {
         }
         const tree = this.props.labelTree
         if (tree) {
-            const graph = mapTreeToGraph(tree)
             return (
                 <React.Fragment>
                     {this.renderEditLabel(tree)}
-
+                    <LabelsPage
+                        labelTree={this.props.labelTree}
+                        visLevel={this.props.visLevel}
+                    ></LabelsPage>
                     {/* <Graph  graph={graph} options={options} events={events} /> */}
                 </React.Fragment>
             )